+++ usblib/host/usbhaudio.h
@@ -2,7 +2,7 @@
 //
 // usbhaudio.h - USB host audio class driver.
 //
-// Copyright (c) 2010 Texas Instruments Incorporated.  All rights reserved.
+// Copyright (c) 2010-2011 Texas Instruments Incorporated.  All rights reserved.
 // Software License Agreement
 // 
 // Texas Instruments (TI) is supplying this software for use solely and
@@ -18,7 +18,7 @@
 // CIRCUMSTANCES, BE LIABLE FOR SPECIAL, INCIDENTAL, OR CONSEQUENTIAL
 // DAMAGES, FOR ANY REASON WHATSOEVER.
 // 
-// This is part of revision 6594 of the Stellaris USB Library.
+// This is part of revision 7611 of the Stellaris USB Library.
 //
 //*****************************************************************************
 
@@ -103,9 +103,9 @@ typedef struct
 extern unsigned long USBHostAudioOpen(unsigned long ulIndex,
                                       tUSBHostAudioCallback pfnCallback);
 extern void USBHostAudioClose(unsigned long ulInstance);
-extern int USBHostAudioPlay(unsigned long ulInstance, void *pvBuffer,
-                            unsigned long ulSize,
-                            tUSBHostAudioCallback pfnCallback);
+extern long USBHostAudioPlay(unsigned long ulInstance, void *pvBuffer,
+                             unsigned long ulSize,
+                             tUSBHostAudioCallback pfnCallback);
 
 extern unsigned long USBHostAudioFormatGet(unsigned long ulInstance,
                                            unsigned long ulSampleRate,
@@ -118,9 +118,9 @@ extern unsigned long USBHostAudioFormatSet(unsigned long ulInstance,
                                            unsigned long ulChannels,
                                            unsigned long ulFlags);
 
-extern int USBHostAudioRecord(unsigned long ulInstance, void *pvBuffer,
-                              unsigned long ulSize,
-                              tUSBHostAudioCallback);
+extern long USBHostAudioRecord(unsigned long ulInstance, void *pvBuffer,
+                               unsigned long ulSize,
+                               tUSBHostAudioCallback);
 
 extern unsigned long USBHostAudioVolumeGet(unsigned long ulInstance,
                                            unsigned long ulInterface,
